+++ pkg/cloudprovider/aws/utils/utils.go
@@ -12,7 +12,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package awsutils
+package utils
 
 import "github.com/awslabs/karpenter/pkg/apis/provisioning/v1alpha1"
